Add test cases that verify subscription date fields
It seems that TypeScript allows strings to be assigned to Date type
fields, but the underlying value is still treated as a string. I have
added some failing test cases to verify that the type is set
correctly. Code needs to be added to convert these fields when
constructing the subscription objects.